refactor(auth, novu, constants): standardize time types, naming, and payload semantics
- Change time constants from int to long for type safety
- Use snake_case for Novu payload fields (consistency with Novu)
- Add Javadoc comments to all constants
- Replace hardcoded values with constants in VerificationTokenService
- Make error messages dynamic (use constants instead of hardcoded strings)
- Rename isNew to isEmailChange in AppUserVerificationPayload
- Fix logic: newValue != null means email change (was inverted)
- Update JSON field from is_new_user to is_email_change
- Add OTP_LOCKOUT_MINUTES and OTP_LOCKOUT_SECONDS constants

+++ src/main/java/com/mentalhealthforum/mentalhealthforum_backend/controller/auth/AuthController.java
@@ -70,7 +70,7 @@ public class AuthController {
                         return ResponseEntity.ok(new StandardSuccessResponse<>("Login successful."));
                     }
                     else if(result instanceof AuthResult.MfaRequired(
-                            String stateToken, String email, int otpLength, int expirySeconds
+                            String stateToken, String email, int otpLength, long expirySeconds
                     )){
                         cookieUtils.setMfaStateCookie(
                                 httpResponse,
