+++ src/main/java/com/mentalhealthforum/mentalhealthforum_backend/dto/novu/OtpPayload.java
@@ -1,11 +1,10 @@
 package com.mentalhealthforum.mentalhealthforum_backend.dto.novu;
 
-import com.mentalhealthforum.mentalhealthforum_backend.contants.AppConstants;
 import com.mentalhealthforum.mentalhealthforum_backend.service.NovuPayload;
 
 import java.util.Map;
 
-public record OtpPayload(String code, int expiryMinutes) implements NovuPayload {
+public record OtpPayload(String code, long expiryMinutes) implements NovuPayload {
     @Override
     public Map<String, Object> toPayloadMap() {
         return Map.of(
