+++ src/main/java/com/mentalhealthforum/mentalhealthforum_backend/dto/userProfileAndIdentity/auth/MfaChallengeResponse.java
@@ -4,5 +4,5 @@ public record MfaChallengeResponse(
         boolean mfaRequired,
         String email,
         int otpLength,
-        int expirySeconds
+        long expirySeconds
 ) {}
